aggr: move decoding logic to helper class
This makes it easier to reuse with different policies for
validation and rollups. For some future work with sharding
the aggregator data this should be useful to have different
policies after processing the initial payload from the user.
