+++ atlas-aggregator/src/test/scala/com/netflix/atlas/aggregator/UpdateApiSuite.scala
@@ -19,7 +19,6 @@ import akka.http.scaladsl.model.HttpEntity
 import akka.http.scaladsl.model.StatusCode
 import akka.http.scaladsl.model.StatusCodes
 import com.fasterxml.jackson.core.JsonFactory
-import com.netflix.atlas.aggregator.UpdateApi.TagMap
 import com.netflix.atlas.akka.ByteStringInputStream
 import com.netflix.atlas.core.util.SmallHashMap
 import com.netflix.atlas.core.util.Strings
